new commit

+++ src/app/reclamations/reclamations.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { Reclamation } from '../model/reclamation.model';
 import { ReclamationService } from '../services/reclamation.service';
 import { Agent } from '../model/agent.model';
+import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-reclamations',
@@ -10,55 +11,25 @@ import { Agent } from '../model/agent.model';
 })
 export class ReclamationsComponent implements OnInit {
 
+
   reclamations! : Reclamation[]; //un tableau de Reclam
   agents! : Agent[];
-  // objet!: string;
-  // searchTerm!: string;
-  // AllReclamations! : Reclamation[];
- 
   
+  constructor(private reclamationService: ReclamationService,
+    private router : Router) {
   
-
-  constructor(private reclamationService: ReclamationService) {
-  //this.reclamations =  reclamationService.listeReclamations();
- // this.agents = reclamationService.listeAgents();
-    
   
 
 }
 ngOnInit(): void {
-  // this.reclamationService.listeReclamation().subscribe(reclams => {
-  //   console.log(reclams);
-  //   this.AllReclamations = reclams;
-  // });
-
   this.chargerReclamations();
   this.chargerAgents();
 }
 
 
-
-// onKeyUp(event: KeyboardEvent) {
-//   const filterText = (event.target as HTMLInputElement).value.toLowerCase() || '';
-//   if (!filterText) {
-//     this.reclamations = [...this.AllReclamations];
-//     return;
-//   }
-
-//   this.reclamationService.rechercherParObjet(filterText).subscribe(filteredReclamations => {
-//     console.log(filteredReclamations);
-//     this.reclamations = filteredReclamations;
-//   });
-// }
-
-
-
-
 chargerAgents(){ 
   this.reclamationService.listeAgents().subscribe(agt => { console.log(agt); 
     this.agents = agt; });
-   
-
   }
 
 chargerReclamations(){ this.reclamationService.listeReclamation().subscribe(reclam => { 
@@ -67,8 +38,28 @@ chargerReclamations(){ this.reclamationService.listeReclamation().subscribe(recl
 }); 
 }
 
- supprimerReclamation(r: Reclamation) { let conf = confirm("Etes-vous sûr ?");
- if (conf) this.reclamationService.supprimerReclamation(r.reference).subscribe(() => {
-   console.log("reclamation supprimé"); this.chargerReclamations(); }); }
+supprimerReclamation(r: Reclamation) { let conf = confirm("Etes-vous sûr ?");
+  if (conf) this.reclamationService.supprimerReclamation(r.reference).subscribe(() => {
+    console.log("reclamation supprimé"); this.chargerReclamations(); }); }
+
+
+  //  archiverReclamation(r: Reclamation) {
+  //   const conf = confirm("Voulez-vous archiver cette réclamation ?");
+  // if (conf) {
+  //   this.reclamationService.archiveReclamation(r.reference.toString()).subscribe(
+  //     () => {
+  //       console.log("Réclamation archivée");
+  //       // Reload the list of reclamations after archiving
+  //       this.chargerReclamations();
+  //     },
+  //     (error) => {
+  //       console.error("Erreur lors de l'archivage de la réclamation", error);
+  //       // Handle error as needed
+  //     }
+  //   );
+  // }
+  //   }
+   
+   
 }
